"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS - 2 Assignments/m6/ADS-2Week-1/Solution.java
@@ -9,7 +9,7 @@ class PageRank {
 		verticesnumber = digraph.V();
 		ranks = new double[verticesnumber];
 		for (int i = 0; i< verticesnumber; i++){
-			ranks[i] = (1/(verticesnumber));
+			ranks[i] = (1/(double)(verticesnumber));
 			System.out.print(i + " : "+ ranks[i]+"\n");
 		}
 		System.out.println();
